+++ lib/src/clojure_lsp/feature/rename.clj
@@ -239,6 +239,11 @@
       {:error {:code :invalid-params
                :message "Can't rename namespace, client does not support file renames."}}
 
+      (and (= :namespace-definitions (:bucket definition))
+           (not= 1 (count (q/find-all-project-namespace-definitions db (:name definition)))))
+      {:error {:code :invalid-params
+               :message "Can't rename namespace, namespace is defined in multiple files."}}
+
       (and (contains? #{:keyword-definitions :keyword-usages} (:bucket definition))
            (not (:ns definition)))
       {:error {:code :invalid-params