+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - Bump clj-kondo to `2023.07.14-20230717.090255-3`. #1624
   - Fix inconsistencies with `:defined-by->lint-as`.
   - Improve memory usage during cache save, avoiding "Out of memory" exceptions.
+  - Prevent file rename when a namespace is defined in multiple files #1574
 
 - Editor
   - Avoid returning all known keywords on empty keywords completion for performance reasons.